+++ src/common/graphs/index.ts
@@ -1,6 +1,7 @@
 import { SquareGraph } from "./square";
+import { SquareOrthGraph } from "./square-orth";
 import { SnubSquareGraph } from "./snubsquare";
 import { HexTriGraph } from "./hextri";
 import { IGraph } from "./IGraph";
 
-export { IGraph, SquareGraph, SnubSquareGraph, HexTriGraph };
+export { IGraph, SquareGraph, SquareOrthGraph, SnubSquareGraph, HexTriGraph };