+++ src/games/index.ts
@@ -10,6 +10,7 @@ import { VolcanoGame, IVolcanoState } from "./volcano";
 import { MvolcanoGame, IMvolcanoState } from "./mvolcano";
 import { ChaseGame, IChaseState } from "./chase";
 import { AbandeGame, IAbandeState } from "./abande";
+import { CephalopodGame, ICephalopodState } from "./ceph";
 
 export {
     APGamesInformation, GameBase, IAPGameState,
@@ -23,14 +24,15 @@ export {
     MvolcanoGame, IMvolcanoState,
     ChaseGame, IChaseState,
     AbandeGame, IAbandeState,
+    CephalopodGame, ICephalopodState,
 };
 
 const games = new Map<string, typeof AmazonsGame | typeof BlamGame | typeof CannonGame |
                               typeof MchessGame | typeof HomeworldsGame | typeof EntropyGame |
                               typeof VolcanoGame | typeof MvolcanoGame | typeof ChaseGame |
-                              typeof AbandeGame>();
+                              typeof AbandeGame | typeof CephalopodGame>();
 // Manually add each game to the following array
-[AmazonsGame, BlamGame, CannonGame, MchessGame, HomeworldsGame, EntropyGame, VolcanoGame, MvolcanoGame, ChaseGame, AbandeGame].forEach((g) => {
+[AmazonsGame, BlamGame, CannonGame, MchessGame, HomeworldsGame, EntropyGame, VolcanoGame, MvolcanoGame, ChaseGame, AbandeGame, CephalopodGame].forEach((g) => {
     if (games.has(g.gameinfo.uid)) {
         throw new Error("Another game with the UID '" + g.gameinfo.uid + "' has already been used. Duplicates are not allowed.");
     }
@@ -60,6 +62,8 @@ export function GameFactory(game: string, ...args: any[]): GameBase|undefined {
             return new ChaseGame(...args);
         case "abande":
             return new AbandeGame(...args);
+        case "ceph":
+            return new CephalopodGame(...args);
     }
     return;
 }